Second Fire Powerup: HomingDrone
What a challenge this was...

When instantiated and there are any GameObjects with the tag "Enemy", the drone will target one of them and home in!

This was a fun task!

Changes:
EnemyDeath method: Since the Enemy can die in several ways now, I made a method solely for its death. Much cleaner now.

Player invulnerability for 1 second after Damage() is called.